Resources Redesign

+++ src/main/java/net/dahliasolutions/controllers/wiki/WikiFileAPIController.java
@@ -51,7 +51,6 @@ public class WikiFileAPIController {
         // get the image save location and verify directories
         String fileName = StringUtils.cleanPath(imageFile.getOriginalFilename());
         String uploadDir = appServer.getResourceDir()+"/wiki/images";
-        String fileURL = appServer.getResourceURL()+"/wiki/images/"+fileName;
         Path uploadPath = Paths.get(uploadDir);
         if (!Files.exists(uploadPath)) {
             try {
@@ -61,10 +60,20 @@ public class WikiFileAPIController {
             }
         }
 
+        // check for existing file
+        Path filePath = uploadPath.resolve(fileName);
+        if (Files.exists(filePath)) {
+            String[] tokens = fileName.split("\\.(?=[^\\.]+$)");
+            String milli = String.valueOf(System.currentTimeMillis());
+            String suffix = milli.substring(milli.length()-4);
+            fileName = tokens[tokens.length-2]+"-"+suffix+"."+tokens[tokens.length-1];
+        }
+        String fileURL = appServer.getResourceURL()+"/wiki/images/"+fileName;
+        filePath = uploadPath.resolve(fileName);
+
         // try to save file
         try {
             InputStream inputStream = imageFile.getInputStream();
-            Path filePath = uploadPath.resolve(fileName);
             Files.copy(inputStream, filePath, StandardCopyOption.REPLACE_EXISTING);
         } catch (IOException e){
             System.out.println("error saving file");
@@ -78,7 +87,17 @@ public class WikiFileAPIController {
     @GetMapping("/removeimage/{id}")
     public String removeStoredWikiIamge(@PathVariable BigInteger id){
         Optional<WikiImage> storedImage = wikiImageService.findById(id);
-        wikiImageService.deleteById(id);
+        if (storedImage.isPresent()) {
+            String trimmedPath = appServer.getResourceDir().replace("/content", "");
+            Path imagePath = Paths.get(trimmedPath+storedImage.get().getFileLocation());
+            try {
+                Files.deleteIfExists(imagePath);
+                System.out.println("Deleting File");
+            } catch (IOException ioException) {
+                System.out.println("File Not Found");
+            }
+            wikiImageService.deleteById(id);
+        }
         return "";
     }
 }
